requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pytest==8.0.0
httpx==0.26.0
pillow==10.2.0
pydantic==2.6.1
requests==2.31.0 